+++ modules/components/src/main/java/io/github/palexdev/mfxcomponents/window/popups/MFXPopup.java
@@ -212,11 +212,14 @@ public class MFXPopup extends PopupControl implements IMFXPopup {
             .map(Scene::getWindow)
             .map(Window::isShowing)
             .orElse(true);
-        if (!showing) return;
+        if (!showing) {
+            close();
+            return;
+        }
 
         retrieveSkin().ifPresentOrElse(
             IMFXPopupSkin::animateOut,
-            super::hide
+            this::close
         );
     }